Teleport portals now work!

+++ SAUCE/famidash.h
@@ -256,7 +256,7 @@ uint8_t settingvalue;
 
 uint8_t gameState = 0x01;
 
-uint16_t teleport_output;
+uint8_t teleport_output;
 
 uint8_t parallax_scroll_x;
 uint8_t invincible_counter;
